+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/common/AufrufKontextToHelper.java
@@ -16,13 +16,13 @@
  */
 package de.bund.bva.isyfact.serviceapi.common;
 
-import de.bund.bva.isyfact.serviceapi.core.aufrufkontext.AufrufKontextToResolver;
 import de.bund.bva.isyfact.serviceapi.core.aufrufkontext.DefaultAufrufKontextToResolver;
 import de.bund.bva.pliscommon.serviceapi.service.httpinvoker.v1_0_0.AufrufKontextTo;
 
 /**
  * Helperclass to resolve AufrufKontextTo from service parameters.
- * @deprecated use non-static {@link AufrufKontextToResolver} as a bean instead.
+ * @deprecated This module is deprecated and will be removed in a future release.
+ * It is recommended to use <a href="https://isyfact.github.io/isyfact-standards-doku/latest/isy-service-rest/konzept/master.html">REST according to IsyFacts REST Concept</a> instead.
  */
 @Deprecated
 public class AufrufKontextToHelper {