+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/core/httpinvoker/TimeoutWiederholungHttpInvokerRequestExecutor.java
@@ -35,7 +35,10 @@ import de.bund.bva.isyfact.serviceapi.common.konstanten.EreignisSchluessel;
 /**
  * Extension to the Spring {@link SimpleHttpInvokerRequestExecutor} which allows
  * to configure the timeout and the times a call is repeated.
+ * @deprecated This module is deprecated and will be removed in a future release.
+ * It is recommended to use <a href="https://isyfact.github.io/isyfact-standards-doku/latest/isy-service-rest/konzept/master.html">REST according to IsyFacts REST Concept</a> instead.
  */
+@Deprecated
 public class TimeoutWiederholungHttpInvokerRequestExecutor extends SimpleHttpInvokerRequestExecutor {
 
     /** Isy-Logger. */