+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/core/serviceimpl/MethodMappingSource.java
@@ -20,8 +20,10 @@ import java.lang.reflect.Method;
 
 /**
  * Diese Schnittstelle bietet Operationen zum Ermitteln einer Zielmethode zu einer aufgerufenen Methode.
- * 
+ * @deprecated This module is deprecated and will be removed in a future release.
+ * It is recommended to use <a href="https://isyfact.github.io/isyfact-standards-doku/latest/isy-service-rest/konzept/master.html">REST according to IsyFacts REST Concept</a> instead.
  */
+@Deprecated
 public interface MethodMappingSource {
 
     /**