+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/core/serviceimpl/ReflectiveExceptionMappingSource.java
@@ -33,7 +33,10 @@ import de.bund.bva.pliscommon.exception.service.PlisToException;
  * Furthermore, it expects that in the RemoteBean operation exactly one TechnicalToException
  * is declared in the RemoteBean operation, which acts as a generic technical exception.
  * </p>
+ * @deprecated This module is deprecated and will be removed in a future release.
+ * It is recommended to use <a href="https://isyfact.github.io/isyfact-standards-doku/latest/isy-service-rest/konzept/master.html">REST according to IsyFacts REST Concept</a> instead.
  */
+@Deprecated
 public class ReflectiveExceptionMappingSource implements ExceptionMappingSource {
 
     @Override