+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/core/serviceimpl/ReflectiveMethodMappingSource.java
@@ -29,7 +29,10 @@ import de.bund.bva.pliscommon.serviceapi.service.httpinvoker.v1_0_0.AufrufKontex
 
 /**
  * Implementation of {@link MethodMappingSource}, which uses reflection to determine the appropriate target method.
+ * @deprecated This module is deprecated and will be removed in a future release.
+ * It is recommended to use <a href="https://isyfact.github.io/isyfact-standards-doku/latest/isy-service-rest/konzept/master.html">REST according to IsyFacts REST Concept</a> instead.
  */
+@Deprecated
 public class ReflectiveMethodMappingSource implements MethodMappingSource {
 
     /**