+++ isy-serviceapi-core/src/main/java/de/bund/bva/isyfact/serviceapi/core/serviceimpl/Validatable.java
@@ -18,8 +18,10 @@ package de.bund.bva.isyfact.serviceapi.core.serviceimpl;
 
 /**
  * Schnittstelle für Implementierungsteile, die ihre statische Konfiguration validieren können.
- * 
+ * @deprecated This module is deprecated and will be removed in a future release.
+ * It is recommended to use <a href="https://isyfact.github.io/isyfact-standards-doku/latest/isy-service-rest/konzept/master.html">REST according to IsyFacts REST Concept</a> instead.
  */
+@Deprecated
 public interface Validatable {
 
     /**